+++ src/lib/libffi/target.mk
@@ -1,4 +1,4 @@
-PKG_DIR = $(call select_from_ports,libffi)/src/lib/gcc/libffi
+PKG_DIR = $(call select_from_ports,libgo)/src/lib/gcc/libffi
 LD_OPT_NOSTDLIB := -nostdlib -Wl,-nostdlib
 MAKE_TARGET := all
 include $(call select_from_repositories,lib/mk/libc-common.inc)
@@ -28,16 +28,16 @@ installed_tar.tag: installed.tag
 	rm -f $(BUILD_BASE_DIR)/debug/libffi.la; \
 	rm -rf $(BUILD_BASE_DIR)/var/libcache/libffi; \
 	if test -e $(BUILD_BASE_DIR)/lib/libffi/.libs/libffi.a; then \
-	echo ".... remove built.tag and installed.tag ....."; \
-	rm $(BUILD_BASE_DIR)/lib/libffi/built.tag; \
-	rm $(BUILD_BASE_DIR)/lib/libffi/installed.tag; \
-	echo ".... remove install dir ....."; \
-	rm -rf $(BUILD_BASE_DIR)/lib/libffi/install; \
-	echo ".... make symlink to ./var/libcache ....."; \
-	mkdir -p $(BUILD_BASE_DIR)/var/libcache/libffi/include/contrib; \
-	ln -sf $(BUILD_BASE_DIR)/lib/libffi/.libs/libffi.a $(BUILD_BASE_DIR)/var/libcache/libffi/libffi.a; \
-	ln -sf $(BUILD_BASE_DIR)/lib/libffi/libffi.la $(BUILD_BASE_DIR)/var/libcache/libffi/libffi.la; \
-	echo ".... copy h-files for to var/libcache/libffi/include ....."; \
-	find $(BUILD_BASE_DIR)/lib/libffi/include/ -name '*.*' -exec cp -fLp {} $(BUILD_BASE_DIR)/var/libcache/libffi/include \; ; \
-	find $(PKG_DIR)/ -name '*.h' -exec cp -fLp {} $(BUILD_BASE_DIR)/var/libcache/libffi/include/contrib \; ; \
+		echo ".... remove built.tag and installed.tag ....."; \
+		rm $(BUILD_BASE_DIR)/lib/libffi/built.tag; \
+		rm $(BUILD_BASE_DIR)/lib/libffi/installed.tag; \
+		echo ".... remove install dir ....."; \
+		rm -rf $(BUILD_BASE_DIR)/lib/libffi/install; \
+		echo ".... make symlink to ./var/libcache ....."; \
+		mkdir -p $(BUILD_BASE_DIR)/var/libcache/libffi/include/contrib; \
+		ln -sf $(BUILD_BASE_DIR)/lib/libffi/.libs/libffi.a $(BUILD_BASE_DIR)/var/libcache/libffi/libffi.a; \
+		ln -sf $(BUILD_BASE_DIR)/lib/libffi/libffi.la $(BUILD_BASE_DIR)/var/libcache/libffi/libffi.la; \
+		echo ".... copy h-files for to var/libcache/libffi/include ....."; \
+		find $(BUILD_BASE_DIR)/lib/libffi/include/ -name '*.*' -exec cp -fLp {} $(BUILD_BASE_DIR)/var/libcache/libffi/include \; ; \
+		find $(PKG_DIR)/ -name '*.h' -exec cp -fLp {} $(BUILD_BASE_DIR)/var/libcache/libffi/include/contrib \; ; \
 	fi
